ElementInfo module changes
Updated the `modules/elementInfo.js` module to clean it up and allow it
to work within the new build system. In addition, some new features were
added:

1. A new, generic `getElement(name, method)` method is provided that
returns an element
2. `getElement.publish`, `getElement.getName`, and `getElement.getID`
are all functions that extend the base `getElement` method.
3. `getElementValue(name)`, `getElementName(name)`, `getElementID(name)`
all use `getElement.publish`, `getElement.getName`, and
`getElement.getID` under the hood respectively.

This module is now included in the library.