Context change
Updated the `context` export to more closely match previous
functionality, as it's simplified form was causing issues when using the
library within page layouts. More work needs to be done to identify the
cause of the issue and remedy it to allow for use within content types
and page layouts.